Added support for directly calling a .md file
This is a basic implementation and only support one template file in which the .md file will be included (/_template.html).

I also need to add some more security test because sanity checks on template paths is really poor right now and this can be exploited from the outsite by just adding ".." in the URL. Not production-ready...

Tests have also been renamed a little bit to try simplify the tests directory.